--- src/deli/base/__init__.py
@@ -1,12 +1,0 @@
-"""base __init__"""
-
-from .selection import BaseSelection, DecodingExperiment
-from .settings import BaseSettings, DecodingSettings
-
-
-__all__ = [
-    "BaseSelection",
-    "BaseSettings",
-    "DecodingExperiment",
-    "DecodingSettings",
-]
