--- src/deli/base/settings.py
@@ -1,174 +1,0 @@
-"""base classes for setting objects"""
-
-from os import PathLike
-from typing import Literal, Self
-
-import yaml
-
-
-class BaseSettings:
-    """
-    Base class for all setting objects
-
-    Setting object are used to organize the possible
-    settings that DELi CLI commands might use.
-
-    DELi uses YAML syntax for all setting files and
-    only supports safe loading of YAML files.
-    This means all variables saved in settings
-    must be primitive objects in python
-    """
-
-    def __init__(self, **kwargs):
-        for key, value in kwargs.items():
-            setattr(self, key, value)
-
-    def to_file(self, path: str | PathLike):
-        """
-        Save settings to a YAML file
-
-        Parameters
-        ----------
-        path : str | PathLike
-            path to save settings to
-        """
-        yaml.dump(self.__dict__, open(path, "w"))
-
-    def get(self, key: str, default: object | None = None) -> object | None:
-        """
-        Get the setting with the matching name
-
-        If a setting of that names does not exist, will return `default`
-
-        Parameters
-        ----------
-        key : str
-            setting name to look up
-        default : object or None
-            the value to return if setting does not exist
-
-        Returns
-        -------
-        object | None
-            the setting value or the default value if setting does not exist
-        """
-        try:
-            return self.__dict__[key]
-        except KeyError:
-            return default
-
-    @classmethod
-    def from_file(cls, path: str | PathLike) -> Self:
-        """
-        Load settings from a yaml file
-
-        Parameters
-        ----------
-        path : str | PathLike
-            Path to yaml file
-
-        Returns
-        -------
-        Self
-        """
-        try:
-            return cls(**yaml.safe_load(open(path, "r")))
-        except Exception as e:
-            raise RuntimeError(f"Failed to load settings from {path}") from e
-
-
-class DecodingSettings(BaseSettings):
-    """
-    Define parameters for decoding experiments
-
-    Only parameters relating to the algorithm should be here
-    Setting relating to IO should be handled outside this context
-    (like in the click command)
-    """
-
-    def __init__(
-        self,
-        library_error_tolerance: float = 0.1,
-        min_library_overlap: int | None = 10,
-        alignment_algorithm: Literal["semi", "hybrid"] = "semi",
-        bb_calling_approach: Literal["alignment", "bio"] = "bio",
-        revcomp: bool = False,
-        max_read_length: int | None = None,
-        min_read_length: int | None = None,
-        read_type: Literal["single", "paired"] = "single",
-        use_hamming: bool = True,
-        umi_clustering: bool = False,
-        umi_min_distance: int = 2,
-    ):
-        """
-        Initialize the decoder settings
-
-        Notes
-        -----
-        More details about the exact effect of these settings can
-        be found in the "Decoding" docs
-
-        Parameters
-        ----------
-        library_error_tolerance: float, default = 0.2
-            the percent error to be tolerated in the library section
-            this will be converted to number of errors based on tag size
-            and down to the nearest whole number
-            for example, a library with 14 nucleotides would tolerate
-            1, 2, and 4 errors for an error tolerance of 0.1, 0.2 and 0.3 respectively
-        min_library_overlap: int or None, default = 7
-            the minimum number of nucleotides required to match
-            the library tag
-            This is because the demultiplexing will accept truncated matches
-            at the front/back of the tag. For example a tag of AGCTGGTTC
-            could match a read of GTTC if the min overlap was <=4
-            If `None`, will default to the exact length of the tag, meaning
-            the whole tag is expected.
-            The recommended value is greater than 8, as the odds of a match this strong
-            to be accidental are low
-        alignment_algorithm: Literal["semi", "hybrid"], default = "semi"
-            the algorithm to use for alignment
-            only used if bb_calling_approach is "alignment"
-        read_type: Literal["single", "paired"], default = "single"
-            the type of read
-            paired are for paired reads
-            all other read types are single
-        revcomp: bool, default = False
-            If true, search the reverse compliment as well
-        max_read_length: int or None, default = None
-            maximum length of a read to be considered for decoding
-            if above the max, decoding will fail
-            if `None` will default to 5x the min_read_length
-        min_read_length: int or None, default = None
-            minimum length of a read to be considered for decoding
-            if below the min, decoding will fail
-            if `None` will default to smallest min match length of
-            any library in the pool considered for decoding
-        bb_calling_approach: Literal["alignment"], default = "alignment"
-            the algorithm to use for bb_calling
-            right now only "alignment" mode is supported
-        use_hamming: bool, default = True
-            enable (`True`) or disable (`False`) hamming decoding
-            only used if a library specifies tags as hamming encoded
-            Note: if hamming encoded libraries are given, and `use_hamming` is
-            `False`, the hamming decoding will not occur, even though it is possible
-        umi_clustering: bool, default = False
-            when doing degeneration, consider two similar UMIs to be the same
-            similarity is based on levenshtein distance and `umi_min_distance`
-        umi_min_distance: int, default = 2
-            the minimum distance between two UMIs to be considered unique
-            only used `umi_clustering` is `True`
-        """
-        super().__init__(
-            library_error_tolerance=library_error_tolerance,
-            min_library_overlap=min_library_overlap,
-            alignment_algorithm=alignment_algorithm,
-            read_type=read_type,
-            revcomp=revcomp,
-            max_read_length=max_read_length,
-            min_read_length=min_read_length,
-            bb_calling_approach=bb_calling_approach,
-            use_hamming=use_hamming,
-            umi_clustering=umi_clustering,
-            umi_min_distance=umi_min_distance,
-        )
