+++ src/deli/decode/__init__.py
@@ -1,13 +1,17 @@
 """init for decoding module"""
 
-from .decoder import DecodedBarcode, DELPoolDecoder, FailedDecode
+from .decoder import DecodedBarcode, DecodeStatistics, FailedDecode
 from .degen import DELibraryPoolCounter, DELibraryPoolIdCounter, DELibraryPoolIdUmiCounter
+from .runner import DecodingExperiment, DecodingExperimentRunner, DecodingSettings
 
 
 __all__ = [
-    "DELPoolDecoder",
     "FailedDecode",
     "DecodedBarcode",
+    "DecodeStatistics",
+    "DecodingExperimentRunner",
+    "DecodingExperiment",
+    "DecodingSettings",
     "DELibraryPoolCounter",
     "DELibraryPoolIdCounter",
     "DELibraryPoolIdUmiCounter",
